Fixed annotations
Reverted (for now) annotations in two helper functions which made mypy trip up. Additionally, changed two assignments (ll. 216 and 504) from "=0" to "=0.0", so that mypy understands we are dealing with floats here. Finally, explicitly annotated variable dx (l. 507) as float, so that mypy doesn't complain when we assign floats to it later (e.g. l. 525).